Factor out processing tab
