+++ vocabsieve/config/word_rules_editor.py
